# Add SharedArrayBuffer example perf test