**Motivation**

[`SharedArrayBuffer`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/SharedArrayBuffer) is another tool in our toolbelt for handling cross-thread communication. It's worth benchmarking a simple example to see what the performance characteristics look like.

**Description**

Add perf test testing sending a `Status` message across a `MessageChannel` vs sending via `SharedArrayBuffer`

Using `SharedArrayBuffer` to pass a `Status` is much faster, seemingly because ssz is faster than `structuredClone`.

```
  SharedArrayBuffer vs MessageChannel
    ✔ set and get status - MessageChannel                              1.268392e+8 ops/s    7.884000 ns/op   x1.039     141832 runs   1.31 s
    ✔ set and get status - SharedArrayBuffer                           5.467469e+8 ops/s    1.829000 ns/op   x1.022     779467 runs   1.82 s
    ✔ serialize/deserialize status - structuredClone                   1.637465e+8 ops/s    6.107000 ns/op        -     168109 runs   1.21 s
    ✔ serialize/deserialize status - ssz                               6.265664e+8 ops/s    1.596000 ns/op        -     684291 runs   1.62 s
```