@dapplion 2023-05-29
lib2p worker PR is merged, and we don't have plans to use ShareArrayBuffer there